+++ src/math-preview.js
@@ -8,11 +8,12 @@ const {jumpToBeginPosition, jumpToEndPosition, getBegin, getEnd} = require('./ut
 
 // init
 let decorationArray = [];
-let macrosArray = [];
+let macrosString = "";
 let macroConfig = vscode.workspace.getConfiguration().get("umath.preview.macros")
 let enablePreview = vscode.workspace.getConfiguration().get('umath.preview.enableMathPreview')
-let showPosition = vscode.workspace.getConfiguration().get('umath.preview.position')
-let renderer = vscode.workspace.getConfiguration().get('umath.preview.renderer')
+let positionConfig = vscode.workspace.getConfiguration().get('umath.preview.position')
+let rendererConfig = vscode.workspace.getConfiguration().get('umath.preview.renderer')
+let cssConfig = vscode.workspace.getConfiguration().get('umath.preview.customCSS')?.join('')
 
 
 // handle MathJax error. Reload on error once.
@@ -23,11 +24,8 @@ let resetError = false;
  * @param {vscode.ExtensionContext} context
  */
 function activate(context) {
-    macrosArray = getMacros(vscode.window?.activeTextEditor?.document, macroConfig)
+    macrosString = getMacros(vscode.window?.activeTextEditor?.document, macroConfig)?.join('\n')??""
     enablePreview = vscode.workspace.getConfiguration().get('umath.preview.enableMathPreview')
-
-    vscode.window.onDidChangeActiveTextEditor((e) => {enablePreview && e && (macrosArray = getMacros(e?.document, macroConfig))})
-    vscode.window.onDidChangeTextEditorSelection((e) => {enablePreview && setPreview(e.textEditor.document, e.selections[0].active)})
 
     context.subscriptions.push(
         vscode.commands.registerCommand('umath.preview.closeAllPreview', clearPreview),
@@ -36,16 +34,18 @@ function activate(context) {
         vscode.commands.registerCommand('umath.preview.reloadPreview', () => {
             if (!enablePreview)
                 return;
-            const editor = vscode.window.activeTextEditor;
-            setPreview(editor.document, editor.selection.active);
+            const editor = vscode?.window?.activeTextEditor;
+            setPreview(editor?.document, editor?.selection?.active);
         }),
-        vscode.workspace.onDidChangeConfiguration(() => {
-            macroConfig = vscode.workspace.getConfiguration().get("umath.preview.macros");
-            showPosition = vscode.workspace.getConfiguration().get('umath.preview.position');
-            renderer = vscode.workspace.getConfiguration().get('umath.preview.renderer');
-            enablePreview = vscode.workspace.getConfiguration().get('umath.preview.enableMathPreview');
-            if(!enablePreview) 
-                clearPreview();
+        vscode.window.onDidChangeActiveTextEditor((e) => {enablePreview && e && (macrosString = getMacros(e?.document, macroConfig)?.join('\n')??"")}),
+        vscode.window.onDidChangeTextEditorSelection((e) => {enablePreview && e && setPreview(e?.textEditor?.document, e?.selections[0]?.active)}),
+        vscode.workspace.onDidChangeConfiguration((e) => {
+            e && e.affectsConfiguration("umath.preview.macros") && (macroConfig = vscode.workspace.getConfiguration().get("umath.preview.macros"));
+            e && e.affectsConfiguration("umath.preview.position") && (positionConfig = vscode.workspace.getConfiguration().get('umath.preview.position'));
+            e && e.affectsConfiguration("umath.preview.renderer") && (rendererConfig = vscode.workspace.getConfiguration().get('umath.preview.renderer'));
+            e && e.affectsConfiguration("umath.preview.enableMathPreview") && (enablePreview = vscode.workspace.getConfiguration().get('umath.preview.enableMathPreview'));
+            e && e.affectsConfiguration("umath.preview.customCSS") && (cssConfig = vscode.workspace.getConfiguration().get('umath.preview.customCSS')?.join(''));
+            !enablePreview && clearPreview();
         })
     )
 }
@@ -93,11 +93,14 @@ function setPreview(document, position) {
     // get rid of "blockquote" & "list"
     if (testScope.isDisplayMath && testScope.scope.indexOf("quote") !== -1)
         mathExpression = mathExpression.replace(/[\n\r]([ \s]*>)+/g, "")
-    mathExpression = (macrosArray?.join('\n')??"") + mathExpression;
-    const previewPosition = (showPosition === 'bottom' && testScope.isDisplayMath 
+
+    // push macros
+    mathExpression = macrosString + mathExpression;
+
+    // set preview position
+    const previewPosition = (positionConfig === 'bottom' && testScope.isDisplayMath 
                         ? new vscode.Position(endInfo.insertPosition.line, endInfo.insertPosition.character - endInfo.match?.matchStr?.length??0) 
                         : beginInfo.insertPosition  )
-    
     pushPreview(mathExpression, testScope.isDisplayMath, previewPosition)
 }
 
@@ -108,13 +111,15 @@ function setPreview(document, position) {
  * @param {vscode.Position} previewPosition 
  */
 function pushPreview(mathExpression, isBlock, previewPosition) {
-    texRenderer[renderer](mathExpression, isBlock)
-        .then((svgString) => {
-            // exclude error case
+    texRenderer[rendererConfig](mathExpression, isBlock)
+    .then((svgString) => {
+        // exclude error case
             if (svgString.includes("error"))
                 return
             // create preview panel
             let mathPreview = createPreview(svgString)
+            // set when clause
+            vscode.commands.executeCommand('setContext','umathShowPreview',true)
             decorationArray.push(mathPreview)
             vscode.window.activeTextEditor.setDecorations(mathPreview, [new vscode.Range(previewPosition, previewPosition)])
         })
@@ -123,14 +128,12 @@ function pushPreview(mathExpression, isBlock, previewPosition) {
             if (onError) {
                 onError = false;
                 resetError = false;
-                return
             }
             else {
                 console.log("retry");
                 onError = true;
-                const editor = vscode.window.activeTextEditor;
-                setPreview(editor.document, editor.selection.active);
-                return
+                const editor = vscode?.window?.activeTextEditor;
+                setPreview(editor?.document, editor?.selection?.active);
             }
         });
 }
@@ -141,31 +144,25 @@ function pushPreview(mathExpression, isBlock, previewPosition) {
  * @returns {vscode.TextEditorDecorationType} 
  */
 function createPreview(mathString) {
-
-    // TODO: add colortheme-kind to global variable
-    const stringColor = (vscode.window.activeColorTheme.kind === vscode.ColorThemeKind.Light ? '#111' : '#fff'); // = Light
-
-    mathString = mathString.replace(/"/g, "'")
-        .replace(/(?<=style\s*=\s*)'/, `'color:${stringColor};`)
-        .replace(/#/g, "%23")
+    const stringColor = ((vscode.window.activeColorTheme.kind === vscode.ColorThemeKind.Dark || vscode.window.activeColorTheme.kind === vscode.ColorThemeKind.HighContrast) ?'#fff' : '#111' ); //  Dark || HC Dark
+    mathString = mathString
+        .replace(/(?<=style\s*=\s*)"/, `"color:${stringColor};`)
+        .split("#").join('%23')
         .replace(/<mjx-container[^<]*><svg/, "<svg")
         .replace("</mjx-container>", "")
 
-    const defaultCss = objectToCssString({
-        // TODO: is there a better way to process svg string? 
-        // Another way to escape \" is : replace(/"/g, "\\\"")
-        ['content']: `url("data:image/svg+xml;utf8,${mathString}")`,
-        ['position']: 'absolute',
-        ['padding']: '0.5em',
+    const defaultCss = 
         // Info: Text and preview SVG are positioned in reverseshow.
-        [showPosition === 'top' ? 'bottom' : 'top']: `1.15rem`,
-        ['display']: `inline-block`,
-        ['z-index']: 1,
-        ['pointer-events']: 'none', //`auto`,
-        ['background-color']: `var(--vscode-editor-background)`,
-        ['border']: '0.5px solid var(--vscode-editorWidget-border)'
-        // TODO: set scale
-    })
+        `content: url('data:image/svg+xml;utf8,${mathString}');\
+        position: absolute;\
+        padding: 0.5em;\
+        ${positionConfig === 'top' ? 'bottom' : 'top'}: 1.15em;\
+        display: inline-block;\
+        z-index: 1;\
+        pointer-events: none;\
+        background-color: var(--vscode-editor-background);\
+        border: 0.5px solid var(--vscode-editorWidget-border);`
+         + cssConfig
 
     return vscode.window.createTextEditorDecorationType({
         before: {
@@ -177,35 +174,24 @@ function createPreview(mathString) {
     })
 }
 
-// from `hoovercj.vscode-power-mode` extension
-function objectToCssString(settings) {
-    let value = ''
-    const cssString = Object.keys(settings)
-        .map((setting) => {
-            value = settings[setting]
-            if (typeof value === 'string' || typeof value === 'number') {
-                return `${setting}: ${value};`
-            }
-        })
-        .join(' ')
-    return cssString
-}
-
 /////////////////////////////////////////////////////////
 
 function clearPreview() {
     for (let thisDeco of decorationArray)
         thisDeco.dispose()
     decorationArray.length = 0;
+    vscode.commands.executeCommand('setContext','umathShowPreview',false)
 }
+
+/////////////////////////////////////////////////////////
 
 function reloadMacros() {
     if (!enablePreview) 
         return;
     macroConfig = vscode.workspace.getConfiguration().get('umath.preview.macros')
-    macrosArray = getMacros(vscode.window.activeTextEditor.document, macroConfig)
-    const editor = vscode.window.activeTextEditor;
-    setPreview(editor.document, editor.selection.active);
+    const editor = vscode.window?.activeTextEditor;
+    macrosString = getMacros(editor?.document, macroConfig)?.join('\n')??""
+    setPreview(editor?.document, editor?.selection?.active);
 }
 
 function toggleMathPreview() {
@@ -214,9 +200,9 @@ function toggleMathPreview() {
     clearPreview();
     if (!enablePreview) 
         return;
-    const editor = vscode.window.activeTextEditor;
-    macrosArray = getMacros(editor.document, macroConfig);
-    setPreview(editor.document, editor.selection.active);
+    const editor = vscode?.window?.activeTextEditor;
+    macrosString = getMacros(editor?.document, macroConfig)?.join('\n')??""
+    setPreview(editor?.document, editor?.selection?.active);
 }
 
 module.exports = {